+++ __test__/inputs/BaseInput.spec.js
@@ -36,8 +36,8 @@ describe("inputs/BaseInput", () => {
         // Must be invalid
         componentNode = TestUtils.mount({ value: "" }, BaseInput, props);
         chai.assert.isNotOk(componentNode.instance().isValid(), "Empty string value must be invalid");
-        chai.assert.equal(componentNode.instance().validationResult().props.children.length, 1, "Empty string value must render one alert");
-        chai.assert.equal(componentNode.instance().validationResult().props.children[0].props.children, "Error Test", "Custom messages shoud be rendered inside alert");
+        chai.assert.equal(componentNode.instance().validationResult().length, 1, "Empty string value must render one alert");
+        chai.assert.equal(componentNode.instance().validationResult()[0].props.children, "Error Test", "Custom messages shoud be rendered inside alert");
         componentNode.unmount();
 
         // Must be invalid
@@ -53,7 +53,7 @@ describe("inputs/BaseInput", () => {
             }
         }, BaseInput, props);
         chai.assert.isNotOk(componentNode.instance().isValid(), "Empty string value must be invalid");
-        chai.assert.equal(componentNode.instance().validationResult().props.children[0].props.children, "Goodbye world", "Arguments must pass successfully");
+        chai.assert.equal(componentNode.instance().validationResult()[0].props.children, "Goodbye world", "Arguments must pass successfully");
         componentNode.unmount();
 
         // Must be exception